fully decoupled algorithms implementation logic from AutoML classes, especially for XGBoost so that AutoML will work even if XGB extension is not in the classpath.
