+++ h2o-automl/build.gradle
@@ -4,12 +4,13 @@ dependencies {
   compile project(":h2o-genmodel")
   compile project(":h2o-core")
   compile project(":h2o-algos")
-  compile project(":h2o-ext-xgboost")
+  compileOnly project(":h2o-ext-xgboost")
 
   // Test dependencies only
   testCompile "junit:junit:${junitVersion}"
   testCompile "com.github.stefanbirkner:system-rules:1.18.0"
   testCompile project(path: ":h2o-core", configuration: "testArchives")
+  testCompile project(":h2o-ext-xgboost")
   testRuntimeOnly project(":${defaultWebserverModule}")
 }
