+++ h2o-automl/src/main/java/ai/h2o/automl/AutoMLBuildSpec.java
@@ -2,17 +2,10 @@ package ai.h2o.automl;
 
 import hex.Model;
 import hex.ScoreKeeper.StoppingMetric;
-import hex.deeplearning.DeepLearningModel.DeepLearningParameters;
-import hex.ensemble.StackedEnsembleModel.StackedEnsembleParameters;
-import hex.glm.GLMModel.GLMParameters;
 import hex.grid.HyperSpaceSearchCriteria;
-import hex.tree.drf.DRFModel.DRFParameters;
-import hex.tree.gbm.GBMModel.GBMParameters;
-import hex.tree.xgboost.XGBoostModel.XGBoostParameters;
 import water.H2O;
 import water.Iced;
 import water.Key;
-import water.exceptions.H2OIllegalArgumentException;
 import water.exceptions.H2OIllegalValueException;
 import water.fvec.Frame;
 import water.util.ArrayUtils;
@@ -290,20 +283,15 @@ public class AutoMLBuildSpec extends Iced {
     }
 
     Model.Parameters getCustomizedDefaults(Algo algo) {
-      if (!_algoParameters.containsKey(algo.name())) _algoParameters.put(algo.name(), defaultParameters(algo));
+      if (!_algoParameters.containsKey(algo.name())) {
+        Model.Parameters defaults = defaultParameters(algo);
+        if (defaults != null) _algoParameters.put(algo.name(), defaults);
+      }
       return _algoParameters.get(algo.name());
     }
 
     private Model.Parameters defaultParameters(Algo algo) {
-      switch (algo) {
-        case DeepLearning: return new DeepLearningParameters();
-        case DRF: return new DRFParameters();
-        case GBM: return new GBMParameters();
-        case GLM: return new GLMParameters();
-        case StackedEnsemble: return new StackedEnsembleParameters();
-        case XGBoost: return new XGBoostParameters();
-        default: throw new H2OIllegalArgumentException("Custom parameters are not supported for "+algo.name()+".");
-      }
+      return algo.enabled() ? ModelingStepsRegistry.defaultParameters(algo.name()) : null;
     }
 
     private void addParameterName(Algo algo, String param) {
@@ -328,8 +316,9 @@ public class AutoMLBuildSpec extends Iced {
     private <V> boolean addParameter(Algo algo, String param, V value) {
       Model.Parameters customParams = getCustomizedDefaults(algo);
       try {
-        if (setField(customParams, param, value, FieldNaming.DEST_HAS_UNDERSCORES)
-                || setField(customParams, param, value, FieldNaming.CONSISTENT)) {
+        if (customParams != null
+                && (setField(customParams, param, value, FieldNaming.DEST_HAS_UNDERSCORES)
+                    || setField(customParams, param, value, FieldNaming.CONSISTENT))) {
           addParameterName(algo, param);
           return true;
         } else {
