+++ h2o-automl/src/main/java/ai/h2o/automl/ModelParametersProvider.java
@@ -1,0 +1,9 @@
+package ai.h2o.automl;
+
+import hex.Model.Parameters;
+
+public interface ModelParametersProvider<P extends Parameters> {
+
+    P newDefaultParameters();
+
+}
